v6 tooltips in the new paradigm
