+++ src/themes/classic/components/index.js
@@ -3,11 +3,13 @@ import Button from './button'
 import Code from './code'
 import Pane from './pane'
 import Select from './select'
+import Tooltip from './tooltip'
 
 export default {
   Alert,
   Button,
   Code,
   Pane,
-  Select
+  Select,
+  Tooltip
 }
