+++ src/themes/classic/components/tooltip.js
@@ -1,0 +1,24 @@
+const baseStyle = {
+  paddingY: 8,
+  paddingX: 16,
+  borderRadius: 'radii.1',
+  elevation: 'shadows.3'
+}
+
+const appearances = {
+  card: {
+    backgroundColor: 'white'
+  },
+  default: {
+    color: 'white',
+    backgroundColor: 'palette.neutral.base'
+  }
+}
+
+const sizes = {}
+
+export default {
+  baseStyle,
+  appearances,
+  sizes
+}
